+++ example/config/events.js
@@ -2,6 +2,7 @@ exports.events = {
     bind: {
         message: {
             send: true,
+            promise: true,
             error: true,
             nativeError: true
         }