+++ example/responses/forbidden.js
@@ -1,0 +1,5 @@
+module.exports = function (req, res) {
+    return function (message) {
+        this.status(403).end(message || 'Access forbidden');
+    };
+};